Added the functionality to store User's Images in Database along with other Form details
I added the server.js file to include routes for the /foundItem and /lostItem forms. This allows users to submit item details and saves the data to the respective collections in the database.

I created foundItem.js and lostItem.js files, which define the Mongoose models for the foundItem and lostItem collections. These models ensure the proper structure and schema for the documents.

I developed HTML files (index.html, foundItem.html, and lostItem.html) to provide forms for users to enter item details. These HTML files were placed in the public directory.

I set up the foundItemUploads and lostItemUploads directories to store uploaded images for their corresponding collections.

Lastly, I updated the package.json file to include the necessary dependencies (express and mongoose).