Integrated lostItem database backend with the detail page under #68
Integrated the created lostItem collection in MongoDB with the Lost Item page in react

After submitting the lost Item form, the server with check whether an image is provided or not ( since that was optional ) if the image is provided that will be stored in lostItemImages directory under the Server folder.

If no image is provided then a default null value is applied on the image field and stored in the database.